+++ hypervisor/Makefile
@@ -190,7 +190,7 @@ C_SRCS += arch/x86/guest/ucode.c
 C_SRCS += arch/x86/guest/pm.c
 C_SRCS += arch/x86/guest/assign.c
 C_SRCS += arch/x86/guest/ept.c
-C_SRCS += arch/x86/guest/io_emul.c
+C_SRCS += arch/x86/guest/vmx_io.c
 C_SRCS += arch/x86/guest/virq.c
 C_SRCS += arch/x86/guest/virtual_cr.c
 C_SRCS += arch/x86/guest/vmcs.c
